% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_stats.R
\name{acceptance_rate}
\alias{acceptance_rate}
\title{Compute acceptance rate}
\usage{
acceptance_rate(a)
}
\arguments{
\item{a}{binary vector with length \code{iter}}
}
\value{
a list containing acceptance rate during the whole chain,
during the first half, and during the second half of the chain
}
\description{
Compute acceptance rate
}
