% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{add.path}
\alias{add.path}
\title{Add MCMC chain path to a contour plot of a density}
\usage{
add.path(p, X, color = "firebrick3", alpha = 1, lines = TRUE)
}
\arguments{
\item{p}{a ggplot object created using \code{plot.surface}}

\item{X}{a matrix with two columns}

\item{color}{line or point color}

\item{alpha}{line or point alpha}

\item{lines}{should the samples be connected with lines?}
}
\value{
a ggplot object
}
\description{
Add MCMC chain path to a contour plot of a density
}
