% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_distributions.R
\name{banana}
\alias{banana}
\title{Log density of the banana distiribution (2d)}
\usage{
banana(x, r = 0.05)
}
\arguments{
\item{x}{a point}

\item{r}{a shape parameter}
}
\value{
log density at \code{x}
}
\description{
Log density of the banana distiribution (2d)
}
