% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/target.R
\name{donut.hamiltonian}
\alias{donut.hamiltonian}
\title{Hamiltonian equations of the donut disribution}
\usage{
donut.hamiltonian(location, momentum, mass)
}
\arguments{
\item{location}{state vector of length 2}

\item{momentum}{momentum vector of length 2}

\item{mass}{mass}
}
\value{
list
}
\description{
Hamiltonian equations of the donut disribution
}
