% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R, R/example_distributions.R
\name{donut}
\alias{donut}
\title{Unnormalized log density of a donut disribution}
\usage{
donut(x)

donut(x)
}
\arguments{
\item{x}{a point}

\item{x}{a point}
}
\value{
unnormalized log density at \code{x}

unnormalized log density at \code{x}
}
\description{
Unnormalized log density of a donut disribution

Unnormalized log density of a donut disribution
}
