% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmc.R
\name{hmc}
\alias{hmc}
\title{Hamiltonian Monte Carlo (HMC) sampling}
\usage{
hmc(log_prob, x0 = c(0, 0), iter = 100, sigma_prop = 1)
}
\arguments{
\item{log_prob}{a function that takes input \code{x} and returns
a value that is proportional to the log probability density at \code{x}}

\item{x0}{initial value}

\item{iter}{number of iterations}
}
\value{
a list
}
\description{
Hamiltonian Monte Carlo (HMC) sampling
}
