% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmc.R
\name{hmc.run}
\alias{hmc.run}
\title{Run one HMC chain}
\usage{
hmc.run(log_prob, x0, iter)
}
\arguments{
\item{log_prob}{a function that takes input \code{x} and returns
a value that is proportional to the log probability density at \code{x}}

\item{x0}{initial value}

\item{iter}{number of iterations}
}
\value{
a list
}
\description{
Run one HMC chain
}
