% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\name{metropolis}
\alias{metropolis}
\title{Metropolis sampling}
\usage{
metropolis(log_prob, x0 = c(0, 0), iter = 100, sigma_prop = 1)
}
\arguments{
\item{log_prob}{a function that takes input \code{x} and returns
a value that is proportional to the log probability density at \code{x}}

\item{x0}{initial value}

\item{iter}{number of iterations}

\item{sigma_prop}{std of Gaussian proposal distribution}
}
\value{
a numeric vector of length \code{iter}
}
\description{
Metropolis sampling
}
