% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metropolis.R
\name{metropolis.run}
\alias{metropolis.run}
\title{Run one metropolis chain}
\usage{
metropolis.run(log_prob, x0, iter, draw.prop)
}
\arguments{
\item{log_prob}{a function that takes input \code{x} and returns
a value that is proportional to the log probability density at \code{x}}

\item{x0}{initial value}

\item{iter}{number of iterations}

\item{draw.prop}{a function that takes input \code{x} and
draws a proposal \code{x_prop}}
}
\value{
a list
}
\description{
Run one metropolis chain
}
