% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_distributions.R
\name{peaks}
\alias{peaks}
\title{Log density of a mixture of three Gaussians (2d)}
\usage{
peaks(x, m1 = c(-2, 2), m2 = c(0, -2), m3 = c(2, 2), S1 = diag(2),
  S2 = diag(2), S3 = diag(2))
}
\arguments{
\item{x}{a point}

\item{m1}{mean vector of component 1}

\item{m2}{mean vector of component 2}

\item{m3}{mean vector of component 3}

\item{S1}{covariance matrix of component 1}

\item{S2}{covariance matrix of component 2}

\item{S3}{covariance matrix of component 3}
}
\value{
log density at \code{x}
}
\description{
Log density of a mixture of three Gaussians (2d)
}
