% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_distributions.R
\name{plot_distribution}
\alias{plot_distribution}
\title{Create contour plot of a density}
\usage{
plot_distribution(fun = donut, grid = c(-4, -4, 4, 4), N_grid = 100)
}
\arguments{
\item{fun}{function that evaluates (unnormalized) log density}

\item{grid}{grid bouds}

\item{N_grid}{number of grid points}
}
\value{
plot of the density
}
\description{
Create contour plot of a density
}
