% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plt.surface}
\alias{plt.surface}
\title{Create contour plot of a density}
\usage{
plt.surface(
  fun = donut,
  grid_size = 3,
  N_grid = 100,
  color_low = "gray95",
  color_high = "gray50",
  title = NULL,
  font_size = 10
)
}
\arguments{
\item{fun}{function that evaluates (unnormalized) log density}

\item{N_grid}{number of grid points}

\item{color_low}{color scale lowest value color}

\item{color_high}{color scale highest value color}

\item{title}{plot title}

\item{font_size}{title font size}

\item{grid}{grid bouds}
}
\value{
a ggplot object
}
\description{
Create contour plot of a density
}
