% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_stats.R
\name{postproc}
\alias{postproc}
\title{Things to do after running a sampler}
\usage{
postproc(chain)
}
\arguments{
\item{chain}{a list}
}
\value{
an updated version of \code{chain}
}
\description{
Things to do after running a sampler
}
