% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_chain.R
\name{run.chain.metropolis}
\alias{run.chain.metropolis}
\title{Run one metropolis chain}
\usage{
run.chain.metropolis(log_prob, x0, iter, proposal_dist)
}
\arguments{
\item{log_prob}{a function that takes input \code{x} and returns
a value that is proportional to the log probability density at \code{x}}

\item{x0}{initial value}

\item{iter}{number of iterations}

\item{proposal_dist}{a function that takes input \code{x} and
draws a proposal \code{x_prop}}
}
\value{
a matrix of size \code{iter} x \code{length(x0)}
}
\description{
Run one metropolis chain
}
