% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmc.R
\name{hmc.run}
\alias{hmc.run}
\title{Run one HMC chain}
\usage{
hmc.run(log_prob, hamiltonian, x0, iter, step_size)
}
\arguments{
\item{log_prob}{a function that takes input \code{theta} and returns
a value that is proportional to the log probability density at \code{theta}}

\item{hamiltonian}{a function that takes as input the location and momentum
and returns their time derivatives}

\item{x0}{initial value}

\item{iter}{number of iterations}

\item{step_size}{leapfrog step size}
}
\value{
a list
}
\description{
Run one HMC chain
}
